requirements.txt
tensorflow==2.13.0
streamlit==1.28.1
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
Pillow==10.0.1
plotly==5.17.0
reportlab==4.0.4
kaggle==1.5.16
tqdm==4.66.1
